test: split tests into multiple files
