+++ test/api/index.spec.js
@@ -1,0 +1,8 @@
+'use strict'
+
+describe('API', () => {
+  it('has the api object', () => {
+    assert(apiClients['a'])
+    assert(apiClients['a'].id)
+  })
+})
