+++ test/api/mount.spec.js
@@ -1,0 +1,5 @@
+'use strict'
+
+describe('.mount', () => {
+  // requires FUSE to be installed, not practical for testing
+})
