+++ test/api/send.spec.js
@@ -1,0 +1,5 @@
+'use strict'
+
+describe('.send', () => {
+  it('used by every command')
+})
