+++ test/api/update.spec.js
@@ -1,0 +1,7 @@
+'use strict'
+
+describe('.update (currently disabled, wait for IPFS 0.4.0 release', () => {
+  it('.update.apply')
+  it('.update.check')
+  it('.update.log')
+})
